fix: core, dataExtractor and datatable extension
