+++ EMS/core-bundle/src/Twig/DatatableExtension.php
@@ -5,13 +5,16 @@ declare(strict_types=1);
 namespace EMS\CoreBundle\Twig;
 
 use EMS\CoreBundle\Service\DatatableService;
+use Twig\Attribute\AsTwigFunction;
 use Twig\Environment;
-use Twig\Extension\RuntimeExtensionInterface;
 
-final readonly class DatatableRuntime implements RuntimeExtensionInterface
+readonly class DatatableExtension
 {
-    public function __construct(private DatatableService $datatableService, private Environment $twig, private string $templateNamespace)
-    {
+    public function __construct(
+        private DatatableService $datatableService,
+        private Environment $twig,
+        private string $templateNamespace
+    ) {
     }
 
     /**
@@ -19,6 +22,7 @@ final readonly class DatatableRuntime implements RuntimeExtensionInterface
      * @param list<string>         $contentTypeNames
      * @param array<string, mixed> $options
      */
+    #[AsTwigFunction('emsco_datatable', isSafe: ['html'])]
     public function generateDatatable(array $environmentNames, array $contentTypeNames, array $options): string
     {
         $datatable = $this->datatableService->generateDatatable($environmentNames, $contentTypeNames, $options);
@@ -34,9 +38,10 @@ final readonly class DatatableRuntime implements RuntimeExtensionInterface
      * @param string[]             $contentTypeNames
      * @param array<string, mixed> $options
      */
-    public function getExcelPath(array $environmentNames, array $contentTypeNames, array $options): string
+    #[AsTwigFunction('emsco_datatable_csv_path', isSafe: ['html'])]
+    public function getCsvPath(array $environmentNames, array $contentTypeNames, array $options): string
     {
-        return $this->datatableService->getExcelPath($environmentNames, $contentTypeNames, $options);
+        return $this->datatableService->getCsvPath($environmentNames, $contentTypeNames, $options);
     }
 
     /**
@@ -44,8 +49,9 @@ final readonly class DatatableRuntime implements RuntimeExtensionInterface
      * @param string[]             $contentTypeNames
      * @param array<string, mixed> $options
      */
-    public function getCsvPath(array $environmentNames, array $contentTypeNames, array $options): string
+    #[AsTwigFunction(name: 'emsco_datatable_excel_path', isSafe: ['html'])]
+    public function getExcelPath(array $environmentNames, array $contentTypeNames, array $options): string
     {
-        return $this->datatableService->getCsvPath($environmentNames, $contentTypeNames, $options);
+        return $this->datatableService->getExcelPath($environmentNames, $contentTypeNames, $options);
     }
 }
